Refactor PetstoreConsumerTests to allow multiple versions
